requirements.txt
bentoml
joblib==1.2.0
pip-chill==1.0.1
torchvision==0.24.1
tqdm==4.64.1
wincertstore==0.2
torchsummary==1.5.1
matplotlib==3.10.3
opencv-python==4.12.0.88
fastapi
uvicorn
dotenv
starlette

